Migrate to global config
