+++ pom.xml
@@ -84,6 +84,10 @@
     </dependency>
     <dependency>
       <groupId>io.jenkins.plugins</groupId>
+      <artifactId>ionicons-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>io.jenkins.plugins</groupId>
       <artifactId>jakarta-mail-api</artifactId>
     </dependency>
     <dependency>
