+++ src/main/resources/hudson/plugins/emailext/ExtendedEmailGlobalConfiguration/global.properties
@@ -1,0 +1,27 @@
+Extended\ E-mail\ Notification=Extended E-mail Notification
+Default\ user\ e-mail\ suffix=Default user e-mail suffix
+Charset=Charset
+Additional\ accounts=Additional accounts
+Account=Account
+Default\ Content\ Type=Default Content Type
+List\ ID=List ID
+Add\ 'Precedence:\ bulk'\ E-mail\ Header=Add 'Precedence: bulk' E-mail Header
+Default\ Recipients=Default Recipients
+Reply\ To\ List=Reply To List
+Emergency\ reroute=Emergency reroute
+Allowed\ Domains=Allowed Domains
+Excluded\ Recipients=Excluded Recipients
+Default\ Subject=Default Subject
+Maximum\ Attachment\ Size=Maximum Attachment Size
+Default\ Attach\ Build\ Log=Default Attach Build Log
+Default\ Content=Default Content
+Default\ Pre-send\ Script=Default Pre-send Script
+Default\ Post-send\ Script=Default Post-send Script
+Additional\ groovy\ classpath=Additional groovy classpath
+Enable\ Throttling=Enable Throttling
+Enable\ Debug\ Mode=Enable Debug Mode
+Require\ Administrator\ for\ Template\ Testing=Require Administrator for Template Testing
+Enable\ watching\ for\ jobs=Enable watching for jobs
+Allow\ sending\ to\ unregistered\ users=Allow sending to unregistered users
+Default\ Triggers=Default Triggers
+Content\ Token\ Reference=Content Token Reference
