+++ src/main/resources/hudson/plugins/emailext/ExtendedEmailGlobalConfiguration/global_ja.properties
@@ -1,0 +1,7 @@
+Extended\ E-mail\ Notification=\u62e1\u5f35E-mail\u901a\u77e5
+Default\ user\ e-mail\ suffix=\u30c7\u30d5\u30a9\u30eb\u30c8 E-mail\u306e\u30b5\u30d5\u30a3\u30c3\u30af\u30b9
+Charset=\u6587\u5b57\u30bb\u30c3\u30c8
+Default\ Content\ Type=\u30c7\u30d5\u30a9\u30eb\u30c8\u30b3\u30f3\u30c6\u30f3\u30c4\u5f62\u5f0f
+Default\ Subject=\u30c7\u30d5\u30a9\u30eb\u30c8\u30b5\u30d6\u30b8\u30a7\u30af\u30c8
+Default\ Content=\u30c7\u30d5\u30a9\u30eb\u30c8\u30b3\u30f3\u30c6\u30f3\u30c4
+Content\ Token\ Reference=\u30c8\u30fc\u30af\u30f3\u30ea\u30d5\u30a1\u30ec\u30f3\u30b9
