+++ src/main/resources/hudson/plugins/emailext/ExtendedEmailGlobalConfiguration/global_zh_TW.properties
@@ -1,0 +1,18 @@
+# The MIT License
+#
+# Copyright (c) 2013, Chunghwa Telecom Co., Ltd., Pei-Tang Huang
+
+Extended\ E-mail\ Notification=\u64f4\u5145\u96fb\u5b50\u90f5\u4ef6\u901a\u77e5
+Default\ user\ e-mail\ suffix=\u9810\u8a2d\u4f7f\u7528\u8005\u4fe1\u7bb1\u5f8c\u7db4
+Charset=\u5b57\u5143\u96c6
+Default\ Content\ Type=\u9810\u8a2d\u5167\u5bb9\u985e\u578b
+List\ ID=List ID
+Add\ 'Precedence:\ bulk'\ E-mail\ Header=\u52a0\u5165 ''Precedence: bulk'' \u90f5\u4ef6\u6a19\u982d
+Default\ Recipients=\u9810\u8a2d\u6536\u4ef6\u4eba
+Reply\ To\ List=\u56de\u8986\u6e05\u55ae
+Emergency\ reroute=\u7dca\u6025\u90f5\u8def
+Default\ Subject=\u9810\u8a2d\u4e3b\u65e8
+Maximum\ Attachment\ Size=\u9644\u4ef6\u5927\u5c0f\u4e0a\u9650
+Default\ Content=\u9810\u8a2d\u5167\u5bb9
+Enable\ Debug\ Mode=\u555f\u7528 Debug \u6a21\u5f0f
+Content\ Token\ Reference=\u5167\u5bb9 Token \u53c3\u8003
