+++ src/main/resources/hudson/plugins/emailext/ExtendedEmailGlobalConfiguration/help-tokens.properties
