+++ src/main/resources/hudson/plugins/emailext/ExtendedEmailPublisher/global.jelly
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+  <f:section title="${%Extended E-mail Notification}">
+    <f:entry>
+      <p>
+        Extended E-mail Notification settings have moved.
+        <a href="${rootURL}/manage/email-ext">Configure Extended E-mail Notification</a>
+      </p>
+    </f:entry>
+  </f:section>
+</j:jelly>
