--- src/main/resources/hudson/plugins/emailext/ExtendedEmailPublisher/global.properties
@@ -1,2 +1,0 @@
-contentType.plainText=Plain Text (text/plain)
-contentType.html=HTML (text/html)
