--- src/main/resources/hudson/plugins/emailext/ExtendedEmailPublisher/global_ja.properties
@@ -1,19 +1,0 @@
-Extended\ E-mail\ Notification=\u62e1\u5f35E-mail\u901a\u77e5
-Override\ Global\ Settings=E-mail\u901a\u77e5\u306e\u8a2d\u5b9a\u3092\u4e0a\u66f8\u304d
-SMTP\ server=SMTP\u30b5\u30fc\u30d0\u30fc
-Default\ user\ E-mail\ suffix=\u30c7\u30d5\u30a9\u30eb\u30c8 E-mail\u306e\u30b5\u30d5\u30a3\u30c3\u30af\u30b9
-System\ Admin\ E-mail\ Address=\u7ba1\u7406\u8005\u306e\u30e1\u30fc\u30eb\u30a2\u30c9\u30ec\u30b9
-Jenkins\ URL=Jenkins URL
-Use\ SMTP\ Authentication=SMTP\u8a8d\u8a3c
-User\ Name=\u30e6\u30fc\u30b6\u30fc\u540d
-Password=\u30d1\u30b9\u30ef\u30fc\u30c9
-Use\ SSL=SSL
-SMTP\ port=SMTP\u30dd\u30fc\u30c8
-Charset=\u6587\u5b57\u30bb\u30c3\u30c8
-Default\ Content\ Type=\u30c7\u30d5\u30a9\u30eb\u30c8\u30b3\u30f3\u30c6\u30f3\u30c4\u5f62\u5f0f
-Default\ Subject=\u30c7\u30d5\u30a9\u30eb\u30c8\u30b5\u30d6\u30b8\u30a7\u30af\u30c8
-Default\ Content=\u30c7\u30d5\u30a9\u30eb\u30c8\u30b3\u30f3\u30c6\u30f3\u30c4
-Content\ Token\ Reference=\u30c8\u30fc\u30af\u30f3\u30ea\u30d5\u30a1\u30ec\u30f3\u30b9
-Help\ for\ feature\:\ Content\ Token\ Reference=\u9ad8\u5ea6\u306a\u6a5f\u80fd\u306e\u30d8\u30eb\u30d7
-contentType.plainText=\u30d7\u30ec\u30a4\u30f3\u30c6\u30ad\u30b9\u30c8 (text/plain)
-contentType.html=HTML (text/html)
