--- src/main/resources/hudson/plugins/emailext/ExtendedEmailPublisher/global_zh_TW.properties
@@ -1,56 +1,0 @@
-# The MIT License
-#
-# Copyright (c) 2013, Chunghwa Telecom Co., Ltd., Pei-Tang Huang
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-# THE SOFTWARE.
-
-Extended\ E-mail\ Notification=\u64f4\u5145\u96fb\u5b50\u90f5\u4ef6\u901a\u77e5
-Override\ Global\ Settings=\u8986\u5beb\u5168\u57df\u8a2d\u5b9a
-SMTP\ server=SMTP \u4f3a\u670d\u5668
-Default\ user\ E-mail\ suffix=\u9810\u8a2d\u4f7f\u7528\u8005\u4fe1\u7bb1\u5f8c\u7db4
-System\ Admin\ E-mail\ Address=\u7cfb\u7d71\u7ba1\u7406\u54e1\u4fe1\u7bb1
-Jenkins\ URL=Jenkins URL
-Use\ SMTP\ Authentication=\u4f7f\u7528 SMTP \u8a8d\u8b49
-User\ Name=\u4f7f\u7528\u8005\u540d\u7a31
-Password=\u5bc6\u78bc
-Use\ SSL=\u4f7f\u7528 SSL
-Use\ TLS=\u4f7f\u7528 TLS
-SMTP\ port=SMTP \u9023\u63a5\u57e0
-Charset=\u5b57\u5143\u96c6
-
-Default\ Content\ Type=\u9810\u8a2d\u5167\u5bb9\u985e\u578b
-contentType.plainText=\u7d14\u6587\u5b57 (text/plain)
-contentType.html=HTML (text/html)
-
-Use\ List-ID\ E-mail\ Header=\u4f7f\u7528 List-ID \u90f5\u4ef6\u6a19\u982d
-List\ ID=List ID
-
-Add\ 'Precedence\:\ bulk'\ E-mail\ Header=\u52a0\u5165 ''Precedence: bulk'' \u90f5\u4ef6\u6a19\u982d
-
-Default\ Recipients=\u9810\u8a2d\u6536\u4ef6\u4eba
-
-Reply\ To\ List=\u56de\u8986\u6e05\u55ae
-Emergency\ reroute=\u7dca\u6025\u90f5\u8def
-Default\ Subject=\u9810\u8a2d\u4e3b\u65e8
-Maximum\ Attachment\ Size=\u9644\u4ef6\u5927\u5c0f\u4e0a\u9650
-Default\ Content=\u9810\u8a2d\u5167\u5bb9
-Enable\ Debug\ Mode=\u555f\u7528 Debug \u6a21\u5f0f
-Enable\ Security=\u555f\u7528\u5b89\u5168\u6027
-Content\ Token\ Reference=\u5167\u5bb9 Token \u53c3\u8003
-Help\ for\ feature\:\ Content\ Token\ Reference=\u529f\u80fd\u8aaa\u660e: \u5167\u5bb9 Token \u53c3\u8003
