+++ src/main/resources/hudson/plugins/emailext/Messages.properties
@@ -1,5 +1,6 @@
 ExtendedEmailPublisher._is_still_in_progress_ignoring_for_purpo={0} is still in progress; ignoring for purposes of comparison
 ExtendedEmailPublisherDescriptor.DisplayName=Editable Email Notification
+ExtendedEmailManagementLink.Description=Configure global defaults for Extended Email notifications.
 ExtendedEmailPublisherDescriptor.AdminAddress=address not configured yet <nobody>
 
 MatrixTriggerMode.OnlyParent=Trigger only the parent job
