+++ src/test/java/hudson/plugins/emailext/ExtendedEmailPublisherDescriptorTest.java
@@ -2,9 +2,6 @@ package hudson.plugins.emailext;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
-import static org.hamcrest.Matchers.empty;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.oneOf;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.jvnet.hudson.test.JenkinsMatchers.hasKind;
 
@@ -52,18 +49,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.function.BiFunction;
 import jenkins.model.Jenkins;
-import org.htmlunit.ElementNotFoundException;
-import org.htmlunit.html.DomElement;
-import org.htmlunit.html.DomNode;
-import org.htmlunit.html.HtmlButton;
-import org.htmlunit.html.HtmlCheckBoxInput;
-import org.htmlunit.html.HtmlDivision;
-import org.htmlunit.html.HtmlNumberInput;
-import org.htmlunit.html.HtmlPage;
-import org.htmlunit.html.HtmlSelect;
-import org.htmlunit.html.HtmlTextArea;
-import org.htmlunit.html.HtmlTextInput;
-import org.jenkinsci.plugins.scriptsecurity.scripts.ScriptApproval;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.jvnet.hudson.test.Issue;
@@ -84,128 +69,6 @@ class ExtendedEmailPublisherDescriptorTest {
         this.j = j;
     }
 
-    private static void assertTitlePage(HtmlPage page) {
-        String actualTitle = page.getTitleText();
-
-        assertTrue(actualTitle.startsWith("System"), "Should be at the Configure System page");
-    }
-
-    @Test
-    void testGlobalConfigDefaultState() throws Exception {
-        HtmlPage page = j.createWebClient().goTo("configure");
-
-        assertTitlePage(page);
-
-        HtmlTextInput smtpHost = page.getElementByName("_.smtpHost");
-        assertNotNull(smtpHost, "SMTP Server should be present");
-        assertEquals("", smtpHost.getText(), "SMTP Server should be blank by default");
-
-        HtmlNumberInput smtpPort = page.getElementByName("_.smtpPort");
-        assertNotNull(smtpPort, "SMTP Port should be present");
-        assertEquals("25", smtpPort.getText(), "SMTP Port should be 25 by default");
-
-        HtmlTextInput defaultSuffix = page.getElementByName("_.defaultSuffix");
-        assertNotNull(defaultSuffix, "Default suffix should be present");
-        assertEquals("", defaultSuffix.getText(), "Default suffix should be blank by default");
-
-        // default content type select control
-        HtmlSelect contentType = page.getElementByName("_.defaultContentType");
-        assertNotNull(contentType, "Content type selection should be present");
-        assertEquals(
-                "text/plain",
-                contentType.getSelectedOptions().get(0).getValueAttribute(),
-                "Plain text should be selected by default");
-
-        HtmlCheckBoxInput precedenceBulk = page.getElementByName("_.precedenceBulk");
-        assertNotNull(precedenceBulk, "Precedence Bulk should be present");
-        assertFalse(precedenceBulk.isChecked(), "Add precedence bulk should not be checked by default");
-
-        HtmlTextInput defaultRecipients = page.getElementByName("_.defaultRecipients");
-        assertNotNull(defaultRecipients, "Default Recipients should be present");
-        assertEquals("", defaultRecipients.getText(), "Default recipients should be blank by default");
-
-        HtmlTextInput defaultReplyTo = page.getElementByName("_.defaultReplyTo");
-        assertNotNull(defaultReplyTo, "Default Reply-to should be present");
-        assertEquals("", defaultReplyTo.getText(), "Default Reply-To should be blank by default");
-
-        HtmlTextInput emergencyReroute = page.getElementByName("_.emergencyReroute");
-        assertNotNull(emergencyReroute, "Emergency Reroute should be present");
-        assertEquals("", emergencyReroute.getText(), "Emergency Reroute should be blank by default");
-
-        HtmlTextInput allowedDomains = page.getElementByName("_.allowedDomains");
-        assertNotNull(allowedDomains, "Allowed Domains should be present");
-        assertEquals("", allowedDomains.getText(), "Allowed Domains should be blank by default");
-
-        HtmlTextInput excludedRecipients = page.getElementByName("_.excludedCommitters");
-        assertNotNull(excludedRecipients, "Excluded Recipients should be present");
-        assertEquals("", excludedRecipients.getText(), "Excluded Recipients should be blank by default");
-
-        HtmlTextInput defaultSubject = page.getElementByName("_.defaultSubject");
-        assertNotNull(defaultSubject, "Default Subject should be present");
-        assertEquals(
-                "$PROJECT_NAME - Build # $BUILD_NUMBER - $BUILD_STATUS!",
-                defaultSubject.getText(),
-                "Default Subject should be set");
-
-        HtmlNumberInput maxAttachmentSize = page.getElementByName("_.maxAttachmentSizeMb");
-        assertNotNull(maxAttachmentSize, "Max attachment size should be present");
-        assertThat(
-                "Max attachment size should be blank or -1 by default",
-                maxAttachmentSize.getText(),
-                is(oneOf("", "-1")));
-
-        HtmlTextArea defaultContent = page.getElementByName("_.defaultBody");
-        assertNotNull(defaultContent, "Default content should be present");
-        assertEquals(
-                "$PROJECT_NAME - Build # $BUILD_NUMBER - $BUILD_STATUS:\n\nCheck console output at $BUILD_URL to view the results.",
-                defaultContent.getText(),
-                "Default content should be set by default");
-
-        HtmlTextArea defaultPresendScript = page.getElementByName("_.defaultPresendScript");
-        assertNotNull(defaultPresendScript, "Default presend script should be present");
-        assertEquals("", defaultPresendScript.getText(), "Default presend script should be blank by default");
-
-        HtmlTextArea defaultPostsendScript = page.getElementByName("_.defaultPostsendScript");
-        assertNotNull(defaultPostsendScript, "Default postsend script should be present");
-        assertEquals("", defaultPostsendScript.getText(), "Default postsend script should be blank by default");
-
-        HtmlCheckBoxInput debugMode = page.getElementByName("_.debugMode");
-        assertNotNull(debugMode, "Debug mode should be present");
-        assertFalse(debugMode.isChecked(), "Debug mode should not be checked by default");
-
-        HtmlCheckBoxInput adminRequiredForTemplateTesting = page.getElementByName("_.adminRequiredForTemplateTesting");
-        assertNotNull(adminRequiredForTemplateTesting, "Admin required for template testing should be present");
-        assertFalse(
-                adminRequiredForTemplateTesting.isChecked(),
-                "Admin required for template testing should be unchecked by default");
-
-        HtmlCheckBoxInput watchingEnabled = page.getElementByName("_.watchingEnabled");
-        assertNotNull(watchingEnabled, "Watching enable should be present");
-        assertFalse(watchingEnabled.isChecked(), "Watching enable should be unchecked by default");
-
-        HtmlCheckBoxInput allowUnregisteredEnabled = page.getElementByName("_.allowUnregisteredEnabled");
-        assertNotNull(allowUnregisteredEnabled, "Allow unregistered should be present");
-        assertFalse(allowUnregisteredEnabled.isChecked(), "Allow unregistered should be unchecked by default");
-
-        assertThrows(
-                ElementNotFoundException.class,
-                () -> page.getElementByName("defaultClasspath"),
-                "defaultClasspath section should not be present");
-    }
-
-    @Test
-    @Issue("JENKINS-20030")
-    void testGlobalConfigSimpleRoundTrip() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        HtmlPage page = j.createWebClient().goTo("configure");
-        HtmlTextInput defaultRecipients = page.getElementByName("_.defaultRecipients");
-        defaultRecipients.setValue("mickey@disney.com");
-        j.submit(page.getFormByName("config"));
-
-        assertEquals("mickey@disney.com", descriptor.getDefaultRecipients());
-    }
-
     @Test
     @Issue("JENKINS-53898")
     void testDefaultAttachBuildLogDefaultValue() {
@@ -215,201 +78,6 @@ class ExtendedEmailPublisherDescriptorTest {
                 0,
                 descriptor.getDefaultAttachBuildLog(),
                 "Default attach build log should be 0 (Do Not Attach) by default");
-    }
-
-    @Test
-    @Issue("JENKINS-53898")
-    void testDefaultAttachBuildLogRoundTrip() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        HtmlPage page = j.createWebClient().goTo("configure");
-        HtmlSelect defaultAttachBuildLog = page.getElementByName("_.defaultAttachBuildLog");
-        assertNotNull(defaultAttachBuildLog, "Default Attach Build Log select should be present in global config");
-        assertEquals(
-                "0",
-                defaultAttachBuildLog.getSelectedOptions().get(0).getValueAttribute(),
-                "Do Not Attach Build Log should be selected by default");
-
-        // Change to "Attach Build Log" (value=1)
-        defaultAttachBuildLog.setSelectedAttribute("1", true);
-        j.submit(page.getFormByName("config"));
-        assertEquals(
-                1,
-                descriptor.getDefaultAttachBuildLog(),
-                "Descriptor should reflect the selected global default (Attach Build Log)");
-
-        // Change to "Compress and Attach Build Log" (value=2)
-        page = j.createWebClient().goTo("configure");
-        defaultAttachBuildLog = page.getElementByName("_.defaultAttachBuildLog");
-        defaultAttachBuildLog.setSelectedAttribute("2", true);
-        j.submit(page.getFormByName("config"));
-        assertEquals(
-                2,
-                descriptor.getDefaultAttachBuildLog(),
-                "Descriptor should reflect the selected global default (Compress and Attach)");
-    }
-
-    @Test
-    @Issue("JENKINS-63367")
-    void testSmtpPortRetainsSetValue() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        JenkinsRule.WebClient client = j.createWebClient();
-        HtmlPage page = client.goTo("configure");
-        HtmlNumberInput smtpPort = page.getElementByName("_.smtpPort");
-        smtpPort.setValue("587");
-        j.submit(page.getFormByName("config"));
-
-        assertEquals("587", descriptor.getMailAccount().getSmtpPort());
-
-        page = client.goTo("configure");
-        smtpPort = page.getElementByName("_.smtpPort");
-        assertEquals("587", smtpPort.getValue());
-    }
-
-    @Test
-    @Issue("JENKINS-20133")
-    void testPrecedenceBulkSettingRoundTrip() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        HtmlPage page = j.createWebClient().goTo("configure");
-        HtmlCheckBoxInput addPrecedenceBulk = page.getElementByName("_.precedenceBulk");
-        addPrecedenceBulk.setChecked(true);
-        j.submit(page.getFormByName("config"));
-
-        assertTrue(descriptor.getPrecedenceBulk());
-    }
-
-    @Test
-    @Issue("JENKINS-20133")
-    void testListIDRoundTrip() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        HtmlPage page = j.createWebClient().goTo("configure");
-        HtmlTextInput listId = page.getElementByName("_.listId");
-        listId.setValue("hammer");
-
-        j.submit(page.getFormByName("config"));
-
-        assertEquals("hammer", descriptor.getListId());
-    }
-
-    @Test
-    void testAdvancedProperties() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        HtmlPage page = j.createWebClient().goTo("configure");
-        HtmlTextArea advProperties = page.getElementByName("_.advProperties");
-        advProperties.setText("mail.smtp.starttls.enable=true");
-        j.submit(page.getFormByName("config"));
-
-        assertEquals(
-                "mail.smtp.starttls.enable=true", descriptor.getMailAccount().getAdvProperties());
-    }
-
-    @Test
-    void defaultTriggers() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        HtmlPage page = j.createWebClient().goTo("configure");
-
-        assertTitlePage(page);
-
-        List<DomNode> nodes = page.getByXPath(".//button[contains(text(),'Default Triggers')]");
-        assertEquals(1, nodes.size());
-        HtmlButton defaultTriggers = (HtmlButton) nodes.get(0);
-        defaultTriggers.click();
-
-        String[] selectedTriggers = {
-            "hudson.plugins.emailext.plugins.trigger.AbortedTrigger",
-            "hudson.plugins.emailext.plugins.trigger.PreBuildTrigger",
-            "hudson.plugins.emailext.plugins.trigger.FixedTrigger",
-            "hudson.plugins.emailext.plugins.trigger.RegressionTrigger",
-        };
-
-        List<DomNode> failureTrigger =
-                page.getByXPath(".//input[@json='hudson.plugins.emailext.plugins.trigger.FailureTrigger']");
-        assertEquals(1, failureTrigger.size());
-        HtmlCheckBoxInput failureTriggerCheckBox = (HtmlCheckBoxInput) failureTrigger.get(0);
-        assertTrue(failureTriggerCheckBox.isChecked());
-        failureTriggerCheckBox.setChecked(false);
-
-        for (String selectedTrigger : selectedTriggers) {
-            List<DomNode> triggerItems =
-                    page.getByXPath(".//input[@name='_.defaultTriggerIds' and @json='" + selectedTrigger + "']");
-            assertEquals(1, triggerItems.size());
-            HtmlCheckBoxInput checkBox = (HtmlCheckBoxInput) triggerItems.get(0);
-            checkBox.setChecked(true);
-        }
-
-        j.submit(page.getFormByName("config"));
-        assertArrayEquals(selectedTriggers, descriptor.getDefaultTriggerIds().toArray(new String[0]));
-    }
-
-    @Test
-    void groovyClassPath() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        HtmlPage page = j.createWebClient().goTo("configure");
-
-        assertTitlePage(page);
-
-        List<DomElement> nodes = page.getByXPath(
-                ".//div[contains(@class, 'setting-name') and ./text()='Additional groovy classpath'] | .//div[contains(@class, 'jenkins-form-label') and ./text()='Additional groovy classpath']");
-        assertEquals(1, nodes.size());
-        HtmlDivision settingName = (HtmlDivision) nodes.get(0);
-
-        nodes = settingName.getByXPath(
-                "../div[contains(@class, 'setting-name')]/div[@class='repeated-container']/div[@name='defaultClasspath']");
-        assertEquals(0, nodes.size(), "Should not have any class path setup by default");
-
-        HtmlDivision div;
-        HtmlButton addButton;
-        nodes = settingName.getByXPath(
-                "../div[@class='setting-main']/div[@class='repeated-container']/button[normalize-space(string(.)) = 'Add']");
-        assertEquals(1, nodes.size());
-        addButton = (HtmlButton) nodes.get(0);
-        addButton.click();
-
-        nodes = settingName.getByXPath(
-                "../div[@class='setting-main']/div[@class='repeated-container']/div[@name='defaultClasspath']");
-        assertEquals(1, nodes.size());
-        div = (HtmlDivision) nodes.get(0);
-        String divClass = div.getAttribute("class");
-        assertTrue(divClass.contains("first") && divClass.contains("last") && divClass.contains("only"));
-
-        nodes = div.getByXPath(".//input[@name='_.path' and @type='text']");
-        assertEquals(1, nodes.size());
-
-        HtmlTextInput path = (HtmlTextInput) nodes.get(0);
-        path.setText("/path/to/classes");
-
-        addButton.click();
-
-        nodes = settingName.getByXPath(
-                "../div[@class='setting-main']/div[@class='repeated-container']/div[@name='defaultClasspath' and contains(@class, 'last')]");
-        assertEquals(1, nodes.size());
-        div = (HtmlDivision) nodes.get(0);
-        divClass = div.getAttribute("class");
-        assertTrue(divClass.contains("last"));
-        assertFalse(divClass.contains("first") || divClass.contains("only"));
-
-        nodes = div.getByXPath(".//input[@name='_.path' and @type='text']");
-        assertEquals(1, nodes.size());
-        path = (HtmlTextInput) nodes.get(0);
-        path.setText("/other/path/to/classes");
-
-        j.submit(page.getFormByName("config"));
-
-        String[] classpath = {
-            "/path/to/classes", "/other/path/to/classes",
-        };
-
-        assertArrayEquals(
-                classpath,
-                descriptor.getDefaultClasspath().stream()
-                        .map(GroovyScriptPath::getPath)
-                        .toArray(String[]::new));
     }
 
     @Test
@@ -435,9 +103,10 @@ class ExtendedEmailPublisherDescriptorTest {
             assertEquals(0, descriptors.size(), "Global configuration should not be accessible to READ users");
         }
         try (ACLContext c = ACL.as(User.getById(MANAGER, true))) {
-            Collection<Descriptor> descriptors = Functions.getSortedDescriptorsForGlobalConfigUnclassified();
+            Collection<Descriptor> descriptors =
+                    Functions.getSortedDescriptorsForGlobalConfigByDescriptor(ExtendedEmailManagementLink.FILTER);
             Optional<Descriptor> found = descriptors.stream()
-                    .filter(descriptor -> descriptor instanceof ExtendedEmailPublisherDescriptor)
+                    .filter(descriptor -> descriptor instanceof ExtendedEmailGlobalConfiguration)
                     .findFirst();
             assertTrue(found.isPresent(), "Global configuration should be accessible to MANAGE users");
         }
@@ -504,116 +173,6 @@ class ExtendedEmailPublisherDescriptorTest {
         ArgumentCaptor<Run<?, ?>> runCaptor = ArgumentCaptor.forClass(Run.class);
         Mockito.verify(authenticatorProvider, Mockito.atLeast(1))
                 .apply(mailAccountCaptor.capture(), runCaptor.capture());
-    }
-
-    @Test
-    void testFixEmptyAndTrimNormal() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        // add a credential to the GLOBAL scope
-        UsernamePasswordCredentials c = new UsernamePasswordCredentialsImpl(
-                CredentialsScope.GLOBAL, "email-ext", "Username/password for SMTP", "smtpUsername", "password");
-        CredentialsProvider.lookupStores(j.jenkins).iterator().next().addCredentials(Domain.global(), c);
-
-        MailAccount ma = new MailAccount();
-        ma.setAddress("example@example.com");
-        ma.setSmtpHost("smtp.example.com");
-        ma.setSmtpPort("25");
-        ma.setCredentialsId("email-ext");
-
-        descriptor.setMailAccount(ma);
-        descriptor.setDefaultSuffix("@example.com");
-        descriptor.setCharset("UTF-8");
-        descriptor.setEmergencyReroute("emergency@example.com");
-        j.submit(j.createWebClient().goTo("configure").getFormByName("config"));
-
-        assertEquals("example@example.com", descriptor.getMailAccount().getAddress());
-        assertEquals("smtp.example.com", descriptor.getMailAccount().getSmtpHost());
-        assertEquals("25", descriptor.getMailAccount().getSmtpPort());
-        assertEquals("email-ext", descriptor.getMailAccount().getCredentialsId());
-        assertEquals("@example.com", descriptor.getDefaultSuffix());
-        assertEquals("UTF-8", descriptor.getCharset());
-        assertEquals("emergency@example.com", descriptor.getEmergencyReroute());
-    }
-
-    @Test
-    void testFixEmptyAndTrimExtraSpaces() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        // add a credential to the GLOBAL scope
-        UsernamePasswordCredentials c = new UsernamePasswordCredentialsImpl(
-                CredentialsScope.GLOBAL, "email-ext", "Username/password for SMTP", "smtpUsername", "password");
-        CredentialsProvider.lookupStores(j.jenkins).iterator().next().addCredentials(Domain.global(), c);
-
-        MailAccount ma = new MailAccount();
-        ma.setAddress("       example@example.com      ");
-        ma.setSmtpHost("      smtp.example.com      ");
-        ma.setSmtpPort("      25      ");
-        ma.setCredentialsId("     email-ext     ");
-
-        descriptor.setMailAccount(ma);
-        descriptor.setDefaultSuffix("      @example.com      ");
-        descriptor.setCharset("      UTF-8      ");
-        descriptor.setEmergencyReroute("      emergency@example.com      ");
-        j.submit(j.createWebClient().goTo("configure").getFormByName("config"));
-
-        assertEquals("example@example.com", descriptor.getMailAccount().getAddress());
-        assertEquals("smtp.example.com", descriptor.getMailAccount().getSmtpHost());
-        assertEquals("25", descriptor.getMailAccount().getSmtpPort());
-        assertEquals("email-ext", descriptor.getMailAccount().getCredentialsId());
-        assertEquals("@example.com", descriptor.getDefaultSuffix());
-        assertEquals("UTF-8", descriptor.getCharset());
-        assertEquals("emergency@example.com", descriptor.getEmergencyReroute());
-    }
-
-    @Test
-    void testFixEmptyAndTrimEmptyString() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        MailAccount ma = new MailAccount();
-        ma.setAddress("");
-        ma.setSmtpHost("");
-        ma.setSmtpPort("");
-        ma.setCredentialsId("");
-
-        descriptor.setMailAccount(ma);
-        descriptor.setDefaultSuffix("");
-        descriptor.setCharset("");
-        descriptor.setEmergencyReroute("");
-        j.submit(j.createWebClient().goTo("configure").getFormByName("config"));
-
-        assertNull(descriptor.getMailAccount().getAddress());
-        assertNull(descriptor.getMailAccount().getSmtpHost());
-        assertEquals("25", descriptor.getMailAccount().getSmtpPort());
-        assertNull(descriptor.getMailAccount().getCredentialsId());
-        assertNull(descriptor.getDefaultSuffix());
-        assertEquals("UTF-8", descriptor.getCharset());
-        assertEquals("", descriptor.getEmergencyReroute());
-    }
-
-    @Test
-    void testFixEmptyAndTrimNull() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        MailAccount ma = new MailAccount();
-        ma.setAddress(null);
-        ma.setSmtpHost(null);
-        ma.setSmtpPort(null);
-        ma.setCredentialsId(null);
-
-        descriptor.setMailAccount(ma);
-        descriptor.setDefaultSuffix(null);
-        descriptor.setCharset(null);
-        descriptor.setEmergencyReroute(null);
-        j.submit(j.createWebClient().goTo("configure").getFormByName("config"));
-
-        assertNull(descriptor.getMailAccount().getAddress());
-        assertNull(descriptor.getMailAccount().getSmtpHost());
-        assertEquals("25", descriptor.getMailAccount().getSmtpPort());
-        assertNull(descriptor.getMailAccount().getCredentialsId());
-        assertNull(descriptor.getDefaultSuffix());
-        assertEquals("UTF-8", descriptor.getCharset());
-        assertEquals("", descriptor.getEmergencyReroute());
     }
 
     @LocalData
@@ -1297,10 +856,10 @@ class ExtendedEmailPublisherDescriptorTest {
     @Test
     @Issue("JENKINS-41473")
     void testBothContentTypeAvailableInGlobalConfig() {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
+        ExtendedEmailGlobalConfiguration globalConfig =
+                ExtendedEmailGlobalConfiguration.get();
 
-        ListBoxModel items = descriptor.doFillDefaultContentTypeItems();
+        ListBoxModel items = globalConfig.doFillDefaultContentTypeItems();
 
         assertEquals(3, items.size(), "Should have exactly three content type options");
 
@@ -1313,12 +872,11 @@ class ExtendedEmailPublisherDescriptorTest {
     @Test
     @Issue("JENKINS-41473")
     void testGlobalBothContentTypeCanBeSetAndInherited() throws Exception {
-        ExtendedEmailPublisherDescriptor descriptor =
-                j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
-        descriptor.setDefaultContentType("both");
-        descriptor.save();
+        ExtendedEmailGlobalConfiguration globalConfig = ExtendedEmailGlobalConfiguration.get();
+        globalConfig.setDefaultContentType("both");
+        globalConfig.save();
 
-        assertEquals("both", descriptor.getDefaultContentType(), "Global content type should be set to 'both'");
+        assertEquals("both", globalConfig.getDefaultContentType(), "Global content type should be set to 'both'");
 
         descriptor = j.jenkins.getDescriptorByType(ExtendedEmailPublisherDescriptor.class);
         assertEquals(
