+++ src/test/resources/configuration-as-code-upgrade.yml
@@ -1,5 +1,5 @@
 unclassified:
-  email-ext:
+  extendedEmailPublisher:
     defaultSuffix: "@domain.extension"
     addAccounts:
       - address: "foo.bar"
